# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: Veritix
  version: "1"
  summary: Audit datasets for integrity problems.
  description: |
    Veritix runs entirely on your own machine or your own cloud. Dataset
    contents never leave the process except where this API says they do, and
    the one place that serves raw customer data — a finding's offending rows —
    has to be asked for by name, one finding at a time.

    Bind address defaults to loopback. Exposing an instance to a network
    requires an auth token; the server refuses to start otherwise.
  license:
    # Dual licensed: the AGPL to everyone, or commercial terms for anyone who
    # needs terms it cannot give. See LICENSING.md in the repository.
    name: AGPL-3.0-or-later, or a commercial license
    identifier: AGPL-3.0-or-later OR LicenseRef-Veritix-Commercial
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /datasets:
    get:
      tags: [datasets]
      summary: List registered datasets.
      operationId: listDatasets
      responses:
        "200":
          description: Datasets, most recently registered first.
          content:
            application/json:
              schema:
                type: object
                required: [datasets]
                properties:
                  datasets:
                    type: array
                    items: { $ref: "#/components/schemas/Dataset" }
        "401": { $ref: "#/components/responses/Unauthorized" }

    post:
      tags: [datasets]
      summary: Register a dataset, by server path or by upload.
      description: |
        A directory is one dataset, not one dataset per file: cross-file keys
        and joins are checked, so the folder is the unit.

        Registering a path is for an operator who already has the data on the
        server. Uploading is how a business user supplies it from a Windows
        desktop; uploaded files are written under the server's data directory
        and never to a path the request controls.
      operationId: createDataset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [path]
              properties:
                path:
                  type: string
                  description: A file or directory on the server's filesystem.
                  examples: ["/srv/exports/retail-2026-08"]
                name:
                  type: string
                  description: Defaults to the base name of the path.
          multipart/form-data:
            schema:
              type: object
              required: [files]
              properties:
                name: { type: string }
                files:
                  type: array
                  description: The files making up one dataset. Names are taken
                    as base names; directory components are rejected.
                  items: { type: string, format: binary }
      responses:
        "201":
          description: Registered. Registering a path that is already known
            returns the dataset that already exists.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dataset" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "413":
          description: The upload exceeded the configured size limit.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
  /datasets/{datasetId}:
    parameters:
      - $ref: "#/components/parameters/DatasetID"
    get:
      tags: [datasets]
      summary: Fetch one dataset.
      operationId: getDataset
      responses:
        "200":
          description: The dataset.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Dataset" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [datasets]
      summary: Forget a dataset and its run history.
      description: Files the server uploaded are deleted with it. A path an
        operator registered in place is left alone — those bytes are theirs.
      operationId: deleteDataset
      responses:
        "204": { description: Deleted. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Required only when the server was started with an auth
        token. A token is mandatory for any non-loopback bind.
  parameters:
    DatasetID:
      name: datasetId
      in: path
      required: true
      schema: { type: string }
...
  responses:
    BadRequest:
      description: The request was malformed.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Missing or wrong bearer token.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: No such resource.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: What went wrong, in words meant for a person.
    Dataset:
      type: object
      required: [id, name, path, uploaded, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        path:
          type: string
          description: Where the data sits on the server.
        uploaded:
          type: boolean
          description: True when the server holds the bytes itself.
        created_at: { type: string, format: date-time }
...